Makes menu measure work dynamically
If you dynamically change the contents of the dropdown, core-dropdown-overlay
will fail to resize itself to suit because it forgets to remove the size that
gets set on it later.

This change simply removes width/height on the element during measurement and
then restores the value before returning the result.